+++ packages/core/tests/js/object/property-node/output.ts
@@ -1,6 +1,6 @@
 const test = {
-	/*a comment updated*/ // prettier-ignore
+	/** a comment */
 	foo: 1,
 
-	/*aka: bond, james bond*/ james: '007'
+	james: '007'
 };