+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sv
 
+## 0.10.7
+### Patch Changes
+
+
+- fix(add): storybook is back to using `@latest` version ([#833](https://github.com/sveltejs/cli/pull/833))
+
+## 0.10.6
+### Patch Changes
+
+
+- fix(cli): files will be formatted after create ([#827](https://github.com/sveltejs/cli/pull/827))
+
 ## 0.10.5
 ### Patch Changes